+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added MakeTablePath, CopyTable, CopyTables, DescribeTable methods for TableClient
 - Add logging for transactions
 
 ## v0.1.5