+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## v0.2.0
 - Added MakeTablePath, CopyTable, CopyTables, DescribeTable methods for TableClient
 - Add logging for transactions